+++ src/RoadRegistry.BackOffice/Core/ModifyRoadSegmentValidator.cs
@@ -38,9 +38,15 @@ public abstract class ModifyRoadSegmentValidatorBase : AbstractValidator<Message
             .WithProblemCode(ProblemCode.RoadSegment.AccessRestriction.NotValid)
             .When(c => c.AccessRestriction != null, ApplyConditionTo.CurrentValidator);
 
-        RuleFor(c => c.Lanes).NotNull();
-        RuleFor(c => c.Widths).NotNull();
-        RuleFor(c => c.Surfaces).NotNull();
+        RuleFor(c => c.Lanes)
+            .NotEmpty()
+            .WithProblemCode(ProblemCode.RoadSegment.Lanes.HasCountOfZero);
+        RuleFor(c => c.Widths)
+            .NotEmpty()
+            .WithProblemCode(ProblemCode.RoadSegment.Widths.HasCountOfZero);
+        RuleFor(c => c.Surfaces)
+            .NotEmpty()
+            .WithProblemCode(ProblemCode.RoadSegment.Surfaces.HasCountOfZero);
     }
 }
 
@@ -71,10 +77,19 @@ public class ModifyRoadSegmentValidator : ModifyRoadSegmentValidatorBase
         RuleFor(c => c.EndNodeId)
             .GreaterThanOrEqualTo(0)
             .NotEqual(c => c.StartNodeId);
-
-        RuleForEach(c => c.Lanes).NotNull().SetValidator(new RequestedRoadSegmentLaneAttributeValidator());
-        RuleForEach(c => c.Widths).NotNull().SetValidator(new RequestedRoadSegmentWidthAttributeValidator());
-        RuleForEach(c => c.Surfaces).NotNull().SetValidator(new RequestedRoadSegmentSurfaceAttributeValidator());
+        
+        RuleForEach(c => c.Lanes)
+            .NotEmpty()
+            .WithProblemCode(ProblemCode.RoadSegment.Lanes.HasCountOfZero)
+            .SetValidator(new RequestedRoadSegmentLaneAttributeValidator());
+        RuleForEach(c => c.Widths)
+            .NotEmpty()
+            .WithProblemCode(ProblemCode.RoadSegment.Widths.HasCountOfZero)
+            .SetValidator(new RequestedRoadSegmentWidthAttributeValidator());
+        RuleForEach(c => c.Surfaces)
+            .NotEmpty()
+            .WithProblemCode(ProblemCode.RoadSegment.Surfaces.HasCountOfZero)
+            .SetValidator(new RequestedRoadSegmentSurfaceAttributeValidator());
     }
 
     private sealed class ModifyRoadSegmentOutlinedValidator : ModifyRoadSegmentValidatorBase
@@ -98,9 +113,18 @@ public class ModifyRoadSegmentValidator : ModifyRoadSegmentValidatorBase
                 .WithProblemCode(ProblemCode.RoadSegment.Morphology.NotValid)
                 .When(c => c.Morphology != null, ApplyConditionTo.CurrentValidator);
 
-            RuleForEach(c => c.Lanes).NotNull().SetValidator(new RequestedRoadSegmentOutlinedLaneAttributeValidator());
-            RuleForEach(c => c.Widths).NotNull().SetValidator(new RequestedRoadSegmentOutlinedWidthAttributeValidator());
-            RuleForEach(c => c.Surfaces).NotNull().SetValidator(new RequestedRoadSegmentOutlinedSurfaceAttributeValidator());
+            RuleForEach(c => c.Lanes)
+                .NotEmpty()
+                .WithProblemCode(ProblemCode.RoadSegment.Lanes.HasCountOfZero)
+                .SetValidator(new RequestedRoadSegmentOutlinedLaneAttributeValidator());
+            RuleForEach(c => c.Widths)
+                .NotEmpty()
+                .WithProblemCode(ProblemCode.RoadSegment.Widths.HasCountOfZero)
+                .SetValidator(new RequestedRoadSegmentOutlinedWidthAttributeValidator());
+            RuleForEach(c => c.Surfaces)
+                .NotEmpty()
+                .WithProblemCode(ProblemCode.RoadSegment.Surfaces.HasCountOfZero)
+                .SetValidator(new RequestedRoadSegmentOutlinedSurfaceAttributeValidator());
         }
     }
 }